+++ stl/lib/src/primitives/int16.dart
@@ -6,7 +6,9 @@ import 'dart:typed_data';
 /// This inherently bounds the variable within standard memory specifications and rigorously
 /// guarantees that mathematical operators intuitively overflow using C++ style
 /// constraints, exactly mimicking real hardware boundaries, providing complete safety!
-extension type Int16(Int16List _data) {
+extension type Int16
+/// Instantiates a dynamically allocated [Int16] bounds-checked element.
+(Int16List _data) {
   /// Dynamically instantiates a [Int16] value mapped sequentially into memory.
   Int16.from(int value) : _data = Int16List(1)..[0] = value;